now should be easier to work on the output when updating/add tests
